Sphere Room Event created (needs polishing up and modifications)
Created rotation event for sphere room.

Code still needs some optimization and modifications for more than two
materials for sphere room.

Also need to reset material to default one after the event is done.

Next goal:
-4th Area room "Roof collapse" event when you pick up a key/examine an
item
-Enemies and events
-Game Design
-Heartbeat manipulation design and implementation

+++ Test Project/Assets/Scripts/rotateMe.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class rotateMe : MonoBehaviour {
 
-    public int speed;
+    public float speed;
     public bool invert;
     // Use this for initialization
     void Start()
@@ -19,15 +19,15 @@ public class rotateMe : MonoBehaviour {
         if (invert)
         {
             temp.y -= speed;
-            //temp.x -= speed;
-           // temp.y -= speed;
+			//temp.x -= speed;
+            //temp.z -= speed;
 
         }
         else
         {
             temp.y += speed;
             //temp.x += speed;
-            //temp.y += speed;
+            //temp.z += speed;
         }
         transform.rotation = Quaternion.Euler(temp.x, temp.y, temp.z);
     }
